requirements.txt

# Django
Django==3.0
django-environ==0.4.5
psycopg2-binary==2.8.6

# DRF
djangorestframework==3.11.0
django-cors-headers==3.5.0
drf-nested-routers==0.92.0

# Custom Choices
django-model-utils==4.0.0

# Swagger
drf-yasg==1.17.1
flex==6.14.1
swagger-spec-validator==2.7.3

# Development
ipdb

# JWT
djangorestframework_simplejwt==4.6.0

# cpf validators
brazilnum==0.8.8
django-localflavor==3.0.1

# Filters
django-filter==2.4.0